BUG/MINOR: debug: properly use long long instead of long for the thread ID
I changed my mind twice on this one and pushed after the last test with
threads disabled, without re-enabling long long, causing this rightful
build warning.

This needs to be backported if the previous commit ff64d3b027 ("MINOR:
threads: export the POSIX thread ID in panic dumps") is backported as
well.

+++ include/common/hathreads.h
@@ -170,7 +170,7 @@ static inline void ha_set_tid(unsigned int tid)
 	ti = &ha_thread_info[tid];
 }
 
-static inline unsigned long ha_get_pthread_id(unsigned int thr)
+static inline unsigned long long ha_get_pthread_id(unsigned int thr)
 {
 	return 0;
 }
